+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - add keep_infinity compress option
 - add quote_style output option
 - add side_effects compress option
+- add switches compress option
 
 Dropped options: